Implementato: update car con campi preimpostati

+++ laravel-main/database/migrations/2021_06_04_122909_create_cars_table.php
@@ -18,7 +18,7 @@ class CreateCarsTable extends Migration
 
             $table->string('name', 64);
             $table->string('model', 64);
-            $table->float('kw', 6, 2);
+            $table->integer('kw');
 
             $table -> bigInteger('brand_id') -> unsigned() -> index();
